# Conflicts:
#	t09ether/target/m2e-wtp/web-resources/META-INF/maven/com.t09ether/home/pom.properties

+++ t09ether/src/main/java/com/t09ether/home/service/OnlineReviewServiceImpl.java
@@ -21,7 +21,7 @@ public class OnlineReviewServiceImpl implements OnlineReviewService{
 	}
 
 	@Override
-	public int rateAvg(int pro_code) {
+	public double rateAvg(int pro_code) {
 		return dao.rateAvg(pro_code);
 	}
